% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlanticForestMetacommunity-package.R
\docType{package}
\name{AtlanticForestMetacommunity-package}
\alias{AtlanticForestMetacommunity}
\alias{AtlanticForestMetacommunity-package}
\title{AtlanticForestMetacommunity: Gives the functions and data for the Analysis of Pelinson et al. 2021 "Disentangling the multiple drivers of amphibian metacommunity structure in the Atlantic Forest at multiple spatial scales"}
\description{
This package is intended to reproduce the data analysis presented
    by Pelinson et al 2021. "Disentangling the multiple drivers of amphibian metacommunity structure in the Atlantic Forest at multiple spatial scales"
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RodolfoPelinson/AtlanticForestMetacommunity}
  \item Report bugs at \url{https://github.com/RodolfoPelinson/AtlanticForestMetacommunity/issues}
}

}
\author{
\strong{Maintainer}: Rodolfo Pelinson \email{rodolfopelinson@gmail.com} (\href{https://orcid.org/0000-0002-8894-588X}{ORCID})

}
\keyword{internal}
