% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_var_function.R
\name{coef_var}
\alias{coef_var}
\title{Coefficient of Variation}
\usage{
coef_var(x)
}
\arguments{
\item{x}{vector from which coefficient of variation will be computed}
}
\description{
Computes de coefficient of variation of a vector
}
