% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_selection_function.R
\name{forward_selection}
\alias{forward_selection}
\title{Automating the forward selection of variables}
\usage{
forward_selection(Y, X)
}
\arguments{
\item{Y}{A matrix of response variables}

\item{X}{A matrix of predictor variables from  variables will be selected.}
}
\description{
This function takes a matrix of variables and excludes those who are not important in explaining species occurrence.
If no variables are important, it does not exclude any variables. It uses the function forward.sel() from package adespatial.
}
