% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lighten_color_function.R
\name{lighten_color}
\alias{lighten_color}
\title{A function lightens a given color}
\usage{
lighten_color(alpha, color)
}
\arguments{
\item{alpha}{value between 0 and 1 where 0 is white and 1 is 100\% the chosen color.}

\item{color}{The color to be lightened}
}
\description{
This function lightens a given color. It basically mixes a given color with white.
}
