% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_color_function.R
\name{mix_color}
\alias{mix_color}
\title{A function to mix two different colors}
\usage{
mix_color(alpha, color1, color2)
}
\arguments{
\item{alpha}{value between 0 and 1 where 0 is 100\% the first color and 1 is 100\% the second color.}

\item{color1}{First color to be mixed}

\item{color2}{Second color to be mixed}
}
\description{
This function mixes two different colors. It relies on functions of the package colorspace
}
