% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/My_Imagine_function.R
\name{My_Imagine}
\alias{My_Imagine}
\title{Plot presence absence matirx with up to five associated environmental gradients.}
\usage{
My_Imagine(
  comm,
  col = c(0, 1, 2),
  order = TRUE,
  scores = 1,
  fill = TRUE,
  xlab = "Species",
  ylab = "Sites",
  yline = 2,
  xline = 2,
  sitenames = rownames(comm),
  speciesnames = colnames(comm),
  Env1 = NULL,
  Env2 = NULL,
  Env.col_1 = NULL,
  Env.label_1 = NULL,
  Env.col_2 = NULL,
  Env.label_2 = NULL,
  cex.site = 1,
  cex.species = 1,
  top_margin = 2,
  left_margin = 2,
  EigenVal = F,
  box.lwd = 2,
  Env.col_3 = NULL,
  Env.label_3 = NULL,
  Env3 = NULL,
  Env.col_4 = NULL,
  Env.label_4 = NULL,
  Env4 = NULL,
  Env.col_5 = NULL,
  Env.label_5 = NULL,
  Env5 = NULL
)
}
\arguments{
\item{comm}{An incidence matrix.}

\item{col}{The colors of absences, true presences and filled presences, respectively.}

\item{order}{Logical. Should the matrix be order by reciprocal averaging (CA)? Defaut set is to TRUE.}

\item{scores}{Numeric. What axis of the CA should be used for ordination? Defaut is 1.}

\item{fill}{Logical. Should embeded absences be filled?}

\item{xlab}{X axis Label. Defaut is to "Species".}

\item{ylab}{Y axis Label. Defaut is to "Sites".}

\item{yline}{Line position of y label. Defaut is 2.}

\item{xline}{Line position of x label. Defaut is 2.}

\item{sitenames}{A vector of site Names. Defaut is row names.}

\item{speciesnames}{A vector of species Names. Defaut is column names.}

\item{Env1}{The first environmental gradient}

\item{Env2}{The second environmental gradient.}

\item{Env.col_1}{The colors of the first Environmental Gradient. If it is a continuous gradient, this should be a vector with the two most extreme colors. If it is a factor, it should be a vector with one color for each of the levels of the factor.}

\item{Env.label_1}{Name of the first environmental gradient.}

\item{Env.col_2}{The colors of the second Environmental Gradient. If it is a continuous gradient, this should be a vector with the two most extreme colors. If it is a factor, it should be a vector with one color for each of the levels of the factor.}

\item{Env.label_2}{Name of the second environmental gradient.}

\item{cex.site}{The size of the site labels.}

\item{cex.species}{The size of the species labels.}

\item{top_margin}{The space left in the top of the plot. Useful to fit long species names.}

\item{left_margin}{The space left in the left side of the plot. Useful to fit long site names.}

\item{EigenVal}{Should the relative eigenvalue of the ordination axis used be plotted?}

\item{box.lwd}{The width of the plot box.}

\item{Env.col_3}{The colors of the third Environmental Gradient. If it is a continuous gradient, this should be a vector with the two most extreme colors. If it is a factor, it should be a vector with one color for each of the levels of the factor.}

\item{Env.label_3}{Name of the third environmental gradient.}

\item{Env3}{The third environmental gradient.}

\item{Env.col_4}{The colors of the fourth Environmental Gradient. If it is a continuous gradient, this should be a vector with the two most extreme colors. If it is a factor, it should be a vector with one color for each of the levels of the factor.}

\item{Env.label_4}{Name of the fourth environmental gradient.}

\item{Env4}{The fourth environmental gradient.}

\item{Env.col_5}{The colors of the fifth Environmental Gradient. If it is a continuous gradient, this should be a vector with the two most extreme colors. If it is a factor, it should be a vector with one color for each of the levels of the factor.}

\item{Env.label_5}{Name of the fifth environmental gradient.}

\item{Env5}{The fifth environmental gradient.}
}
\description{
This function plots the incidence matrix, ordinated or not, with up to five associated environmental gradients associates.
}
