% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/My_spearman_function.R
\name{My_spearman}
\alias{My_spearman}
\title{Performs Spearman rank correlations or Kruskal-Wallis Rank Sum Test on the First CA Axis.}
\usage{
My_spearman(Y, X, p.adjust.method = "fdr")
}
\arguments{
\item{Y}{An incidence matrix.}

\item{X}{A matrix of gradient variables from, either continuous gradients or a factor.}

\item{p.adjust.method}{Method of p-value correction. Same as for the "p.adjust" function.}
}
\description{
This function performs Spearman rank correlations between the first axis of a CA computed on a incidencematrix and multiple gradient variables.
. If the gradient variable is a factor, it performs Kruskal-Wallis Rank Sum Tests. P values are corrected for multiple comparisons.
}
