% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_partitioning_function.R
\name{var_partitioning}
\alias{var_partitioning}
\title{Performing variation partitioning with p values}
\usage{
var_partitioning(
  Y,
  env,
  clim = NULL,
  spa,
  percent_r2 = F,
  allow_negative_r2 = TRUE,
  round = 3
)
}
\arguments{
\item{Y}{Species occurence matrix}

\item{env}{Matrix of environmental variables}

\item{clim}{Matrix of climate variables}

\item{spa}{Matrix of spatial variables}

\item{percent_r2}{Logical. Should the R2 values be converted in percentages?}

\item{allow_negative_r2}{Logical. Should negative values be used to compute the R2 values?}
}
\description{
This function uses the \code{varpart()} function from vegan to compute all fractions and their respective p-values in a single table
}
