% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_partitioning_1_function.R
\name{var_partitioning_1}
\alias{var_partitioning_1}
\title{Performing variation partitioning with p values}
\usage{
var_partitioning_1(Y, env, clim, spa, percent_r2 = F)
}
\arguments{
\item{Y}{Species occurence matrix}

\item{env}{Matrix of environmental variables}

\item{clim}{Matrix of climate variables}

\item{spa}{Matrix of spatial variables}

\item{percent_r2}{Logical. Should the R2 values be converted in percentages?}
}
\description{
This function uses the \code{varpart()} function from vegan to compute all fractions and their respective p-values in a single table
}
