% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_partitioning_2_function.R
\name{var_partitioning_2}
\alias{var_partitioning_2}
\title{Performing variation partitioning with p values}
\usage{
var_partitioning_2(Y, env, spa, allow_negative_r2 = FALSE, percent_r2 = FALSE)
}
\arguments{
\item{Y}{Species occurence matrix}

\item{env}{Matrix of environmental variables}

\item{spa}{Matrix of spatial variables}

\item{allow_negative_r2}{Logical. Should negative values be used to compute the R2 values?}

\item{percent_r2}{Logical. Should the R2 values be converted in percentages?}

\item{clim}{Matrix of climate variables}
}
\description{
This function uses the \code{varpart()} function from vegan to compute all fractions and their respective p-values in a single table. It differs from the \code{var_partitioning_1} because it can or cannot omit negative values in the calculations of R2.
}
