% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIF_selection_function.R
\name{VIF_selection}
\alias{VIF_selection}
\title{Automating the exclusion of multicolinear variables}
\usage{
VIF_selection(Y, X)
}
\arguments{
\item{Y}{A matrix of response variables}

\item{X}{A matrix of predictor variables from which multicolinear variables will be excluded.}
}
\description{
This function takes a matrix of variables and excludes those who exhibit
a variance inflation factor higher than 3 for giving matrix of response variables. It uses the function vif.cca() from package vegan.
}
